Revert "Fixed a bunch of (mostly minor) issues identified by analysis.apache.org"
This reverts commit 08bce28233a03e556b6149f75919ac76ff601969.

git-svn-id: https://svn.apache.org/repos/asf/cxf/branches/2.6.x-fixes@1393523 13f79535-47bb-0310-9956-ffa450edef68

+++ api/src/main/java/org/apache/cxf/common/commands/ForkedCommand.java
@@ -159,7 +159,7 @@ public class ForkedCommand extends Thread {
                 if (LOG.isLoggable(Level.FINE)) {
                     LOG.fine("Waiting " + timeout + " seconds for process to complete");
                 }
-                join((long)timeout * 1000L);
+                join(timeout * 1000);
             } else {
                 if (LOG.isLoggable(Level.FINE)) {
                     LOG.fine("Waiting for process to complete");
