+++ rt/management-web/src/main/java/org/apache/cxf/management/web/logging/atom/deliverer/RetryingDeliverer.java
@@ -81,7 +81,7 @@ public final class RetryingDeliverer implements Deliverer {
             cal = Calendar.getInstance();
             cal.add(Calendar.SECOND, sleep);
             if (timeout == 0 || timeoutDate.after(cal.getTime())) {
-                Thread.sleep((long)sleep * 1000L);
+                Thread.sleep(sleep * 1000);
             } else {
                 pauser.reset();
                 return false;
